# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CargoDTO:
    properties:
      id_cargo:
        type: integer
      nombre_cargo:
        type: string
    type: object
  dto.ContactUpdateRequest:
    properties:
      celular:
        type: string
      email:
        type: string
    required:
    - celular
    - email
    type: object
...
  dto.CreateMedidasRequest:
    properties:
      cadera_cm:
        type: number
      cintura_cm:
        type: number
      estatura_m:
        type: number
      fecha_inicio:
        type: string
      manga_cm:
        type: number
      pecho_cm:
        type: number
    required:
    - estatura_m
    type: object
...
  dto.EstadoDTO:
    properties:
      id_estado:
        type: integer
      nombre_estado:
        type: string
      tabla_estado:
        type: string
    type: object
...
  dto.MedidasResponse:
    properties:
      cadera_cm:
        type: number
      cintura_cm:
        type: number
      estatura_m:
        type: number
      fecha_fin:
        type: string
      fecha_inicio:
        type: string
      id_funcionario:
        type: integer
      id_medidas:
        type: integer
      manga_cm:
        type: number
      pecho_cm:
        type: number
    type: object
...
  dto.ProfileResponse:
    properties:
      apellido_materno:
        type: string
      apellido_paterno:
        type: string
      cargo:
        $ref: '#/definitions/dto.CargoDTO'
      celular:
        type: string
      email:
        type: string
      id:
        type: integer
      medidas:
        $ref: '#/definitions/dto.MedidasResponse'
      nombres:
        type: string
      rut:
        type: string
      sucursal:
        $ref: '#/definitions/dto.SucursalDTO'
      tallas_registradas:
        type: boolean
    type: object
...
  dto.SucursalDTO:
    properties:
      direccion:
        type: string
      estado:
        $ref: '#/definitions/dto.EstadoDTO'
      estado_sucursal:
        type: integer
      id_sucursal:
        type: integer
      nombre_sucursal:
        type: string
      total_funcionarios:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: soporte@sistal.cl
    name: Soporte API
  description: API RESTful para la gestión integral de funcionarios en el sistema
    de control de uniformes empresariales
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: API Microservicio de Funcionarios
  version: "1.0"
paths:
...
  /api/v1/employees/{userId}/contact:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID o 'me'
        in: path
        name: userId
        required: true
        type: string
      - description: Datos de contacto
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ContactUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
      security:
      - BearerAuth: []
      summary: Actualizar contacto
      tags:
      - employees
  /api/v1/employees/{userId}/measurements:
    get:
      parameters:
      - description: User ID o 'me'
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MedidasResponse'
      security:
      - BearerAuth: []
      summary: Obtener medidas
      tags:
      - employees
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID o 'me'
        in: path
        name: userId
        required: true
        type: string
      - description: Medidas
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMedidasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
      security:
      - BearerAuth: []
      summary: Actualizar medidas
      tags:
      - employees
  /api/v1/employees/{userId}/profile:
    get:
      description: Obtiene los datos del funcionario por User ID (o "me")
      parameters:
      - description: User ID o 'me'
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener perfil
      tags:
      - employees
...
  /api/v1/funcionarios/me/stats:
    get:
      description: Obtiene las estadísticas del funcionario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener estadísticas
      tags:
      - funcionarios
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Autenticación mediante token JWT. Formato: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...